+++ Content.Client/Nutrition/EntitySystems/OpenableSystem.cs
@@ -1,0 +1,7 @@
+using Content.Shared.Nutrition.EntitySystems;
+
+namespace Content.Client.Nutrition.EntitySystems;
+
+public sealed class OpenableSystem : SharedOpenableSystem
+{
+}
